backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
aiofiles>=24.1.0
httpx>=0.26.0
pyyaml>=6.0
